Create Forms CSS
This code help to update the Forms CSS to make it look better. I have formatted the forms to show placeholder text instead of label text, then I updated colors, backgrounds, lines, and action button.

For the case of the Popup forms I decided to center it since I think it looks better.